Refactored code to merge style repo and functional repo
