+++ public/index.html
@@ -7,9 +7,14 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
   <meta name="description" content="Web site created using create-react-app" />
-  <style>
-    @import url('https://fonts.googleapis.com/css2?family=Aboreto&family=Prompt:ital,wght@1,500&family=Zen+Dots&display=swap');
-  </style>
+  
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Rock+Salt&display=swap" rel="stylesheet">
+  
+  <!-- <style>
+    @import url(‘https://fonts.googleapis.com/css2?family=Rock+Salt&display=swap’);
+  </style> -->
 
 
   <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
@@ -27,12 +32,13 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Battle of the Minds</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
-    <!--
+  <title>Battle of the Minds</title>
+</head>
+
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="root"></div>
+  <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
