--- src/About.module.css
@@ -1,19 +1,0 @@
-.title {
-  
-}
-
-.teamMemberContainer {
-  
-}
-
-.memberName {
-  
-}
-
-.memberPicture {
-  width: 300px;
-}
-
-.memberBio {
-  
-}
