+++ src/About/About.module.css
@@ -1,0 +1,34 @@
+.container {
+  margin-left: auto;
+  margin-right: auto;
+  text-align: center;
+  font-family: 'Rock Salt', cursive;
+  background-color:#EAE0D5;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.title {
+  padding-top: 25px;
+  padding-bottom: 40px;
+  font-weight: bolder;
+}
+
+.teamMemberContainer {
+  width: 45vw;
+}
+
+/* .memberName {
+} */
+
+.memberPicture {
+  width: 300px;
+}
+
+.memberBio {
+  font-family: cursive;
+  font-weight: 600;
+  margin-top: 10px;
+  margin-bottom: 45px;
+}
