--- src/Header/Header.css
@@ -1,46 +1,0 @@
-.h1 {
-  font-family: 'Zen Dots', cursive;
-  font-size: 4.2rem;
-}
-
-header {
-<<<<<<<< HEAD:src/Header/Header.css
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 20px;
-  background-color: #463f3a;
-========
-  border-bottom: solid black 1px;
->>>>>>>> thursdayDev:src/Header.module.css
-}
-
-.login {
-  text-align: center;
-}
-
-.profilePic {
-  width: 50px;
-}
-<<<<<<<< HEAD:src/Header/Header.css
-.userImg {
-  width: 100px;
-  height: 100px;
-}
-========
-
-.upperRow {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  padding: 5px;
-}
-
-.activeLink {
-  display: none;
-}
-
-.link {
-  padding: 2px;
-}
->>>>>>>> thursdayDev:src/Header.module.css
