+++ src/Leaderboard/Leaderboard.module.css
@@ -1,0 +1,22 @@
+.container {
+  padding-top: 25px;
+  padding-bottom: 55px;
+  text-align: center;
+  font-family: 'Rock Salt', cursive;
+  font-weight: bold;
+  background-color: #EAE0D5;
+  height: 90vw;
+}
+
+.userDiv {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  margin: 5px 20px;
+  border: solid #C6AC8F 1px;
+  font-weight: bold;
+}
+
+.userPic {
+  height: 75px;
+}
