+++ src/Logout/Logout.module.css
@@ -1,4 +1,4 @@
-.button-1 {
+.button1 {
   min-width: 100px;
   margin: 10px;
   height: 40px;
@@ -17,13 +17,13 @@
   border: 2px solid #5e503f;
   color: #5e503f;
 }
-.button-1:hover {
+.button1:hover {
   color: #eae0d5;
 }
-.loginButton:hover:after {
+.button1:hover:after {
   height: 100%;
 }
-.button-1:after {
+.button1:after {
   content: '';
   position: absolute;
   z-index: -1;
