+++ src/Question/Question.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from 'react-bootstrap/Button';
+import styles from './Question.module.css';
 
 class Question extends React.Component {
   constructor(props) {
@@ -13,47 +13,43 @@ class Question extends React.Component {
     this.scrambleAnswersArray();
   }
 
-  // removeSpecialCharacters = (string) => {
-  //   const renderString = string.replace(/&lt;/g, '<').replace(/&gt;/g, '>').replace(/&quot;/g, '"').replace(/&#39;/g, `'`).replace(/&amp;/g, '&');
-  //   return renderString;
-  // };
+  removeSpecialCharacters = (string) => {
+    const renderString = string.replace(/&lt;/g, '<').replace(/&gt;/g, '>').replace(/(&quot;|&ldquo;|&rdquo;)/g, '"').replace(/(&#39;|&#039;|&rsquo;)/g, `'`).replace(/&amp;/g, '&').replace(/&eacute;/g, 'é').replace(/&ndash;/g, '-').replace(/&deg;/g, '°');
+    return renderString;
+  };
 
   scrambleAnswersArray = () => {
-    const scrambledAnswersArray = [
-      ...this.props.question.incorrect_answers,
-      this.props.question.correct_answer,
-    ];
+    const scrambledAnswersArray = [...this.props.question.incorrect_answers, this.props.question.correct_answer];
     scrambledAnswersArray.sort(() => 0.5 - Math.random());
     this.setState({
-      answersArray: scrambledAnswersArray,
+      answersArray: scrambledAnswersArray
     });
   };
 
   render() {
     return (
       <>
-        {this.props.renderedQuestionIndex ===
-        this.props.currentQuestionIndex ? (
-          <form onSubmit={() => this.props.handleSubmitAnswer(event)}>
-            <h1>{this.props.question.question}</h1>
+        {
+          this.props.renderedQuestionIndex === this.props.currentQuestionIndex ?
+            <form onSubmit={() => (this.props.handleSubmitAnswer(event))}>
+              <h1>{this.removeSpecialCharacters(this.props.question.question)}</h1>
 
-            {this.state.answersArray.map((answer, index) => (
-              <label key={index}>
-                <input
-                  name={this.props.question.question}
-                  type='radio'
-                  value={answer}
-                  onChange={(e) =>
-                    this.props.handleAnswerChange(e.target.value)
-                  }
-                />
-                {answer}
-              </label>
-            ))}
-            <br />
-            <Button type='submit'>Submit Answer</Button>
-          </form>
-        ) : null}
+              {this.state.answersArray.map((answer, index) => (
+                <label key={index}>
+
+                  <input
+                    name={this.props.question.question}
+                    type="radio"
+                    value={answer}
+                    onChange={(e) => this.props.handleAnswerChange(e.target.value)} />
+                  {this.removeSpecialCharacters(answer)}
+
+                </label>
+              ))}
+              <br />
+              <button className={styles.startButton} type="submit">Submit Answer</button>
+            </form> : null
+        }
       </>
     );
   }
